If a version is already installed, reinstall with a local eups tag
Basically, we want to be able to reinstall a version after having
updated a dependancy. For example, if tron_actorcore 2.4.11 is
installed and used, and depends on ics_config 1.4.8. We then install
ics_config 1.5.0, and want a new installation of tron_actorcore
2.4.11, but without breaking existing products.

This change will make `pfsinstall tron_actorcore:2.4.11` install that
git version of tron_actorcore, but with a version of `2.4.11a`. We
have been doing this manually when needed.